connectors/postgres.py
import os
import psycopg2
from psycopg2.extras import RealDictCursor
from typing import List, Dict, Any, Optional

class PostgresConnector:
    def __init__(self):
        self.connection = None
        self.cursor = None
    
    def connect(self):
        """Establish connection to PostgreSQL database"""
        # TODO: Add connection parameters from environment variables
        pass
    
    def execute_query(self, query: str, params: Optional[tuple] = None) -> List[Dict[str, Any]]:
        """Execute a SELECT query and return results as list of dicts"""
        # TODO: Implement query execution
        pass
    
    def execute_update(self, query: str, params: Optional[tuple] = None) -> int:
        """Execute an INSERT/UPDATE/DELETE query and return affected rows"""
        # TODO: Implement update execution
        pass
    
    def log_process_stage(self, stage: str, status: str, details: Dict[str, Any]):
        """Log process monitoring information to tracking table"""
        # TODO: Implement process logging
        pass
    
    def close(self):
        """Close database connection"""
        if self.cursor:
            self.cursor.close()
        if self.connection:
            self.connection.close()