requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
certifi>=2023.0.0